Añade soporte para la actualización automática de yt-dlp y el repositorio mediante comandos git. Refactoriza el scheduler para ejecutar descargas periódicas y ajusta la configuración de logging. Actualiza el Dockerfile para incluir git y modifica el archivo .env para establecer la ruta del repositorio.
