+++ Dockerfile
@@ -5,6 +5,7 @@ RUN apt-get update && apt-get install -y \
     ffmpeg \
     curl \
     tzdata \
+    git \
     && rm -rf /var/lib/apt/lists/*
 
 # Instala yt-dlp
