+++ app/utils/logging_config.py
@@ -4,7 +4,7 @@ import sys
 
 def configurar_logging():
     logging.basicConfig(
-        level=logging.DEBUG,  # Cambia a DEBUG si quieres más detalle
+        level=logging.INFO,  # Cambia a DEBUG si quieres más detalle
         format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
         handlers=[logging.StreamHandler(sys.stdout)]
     )
