+++ config/.env
@@ -1,3 +1,5 @@
 ROOT_PATH=/music
 
-SCHEDULE_INTERVAL_DAYS=5
+SCHEDULE_INTERVAL_DAYS=1
+
+REPO_PATH=https://github.com/luishidalgoa/Subscription-Artists-And-Automatic-download-music
